# Improve performance of registry.Register