The use of IncrementCounter vs BatchIncrement counter causes
registry.Registry to be 6~7x slower

with sender.IncrementCounter:
```
go test -run Benchmark -bench=.
BenchmarkRegisterWith100KRoutes-4         100000             23244 ns/op
BenchmarkRegisterWithOneRoute-4           100000             19677 ns/op
PASS
ok      code.cloudfoundry.org/gorouter/registry 17.593s
```

with batcher.BatchIncrementCounter:
```
go test -run Benchmark -bench=.
BenchmarkRegisterWith100KRoutes-4         300000              3616 ns/op
BenchmarkRegisterWithOneRoute-4           500000              3032 ns/op
PASS
ok      code.cloudfoundry.org/gorouter/registry 8.872s
```

This commit changes the behavior of one of Gorouter's metrics:
Instead of the `registry_message` counter metric being emitted on every
NATS message, it will emit the metric on a set interval.

Thanks for contributing to gorouter. To speed up the process of reviewing your pull request please provide us with:

* A short explanation of the proposed change:

* An explanation of the use cases your change solves

* Instructions to functionally test the behavior change using operator interfaces (BOSH manifest, logs, curl, and metrics)

* Expected result after the change

* Current result before the change

* Links to any other associated PRs

* [x] I have viewed signed and have submitted the Contributor License Agreement

* [x] I have made this pull request to the `master` branch

* [x] I have run all the unit tests using `bin/test`

* [ ] I have run CF Acceptance Tests on bosh lite