@shalako 2017-05-15
@flawedmatrix Thank you for helping the team improve performance. 

Will the metric name change? 

How often will it be emitted? Based on other gorouter metrics documented [here](https://docs.cloudfoundry.org/running/all_metrics.html#gorouter), I would expect every 5 seconds.

On merging this PR, we will also need a PR to update the CF docs: https://github.com/cloudfoundry/docs-running-cf/blob/master/metrics/_routing.html.md.erb