@flawedmatrix 2017-05-16
@shalako

The metric name isn't changed as part of this PR. It's emitted on the same interval as all the other batch metrics in gorouter (5 seconds).

Here's a performance report showing the comparison between how gorouter currently handles 100K routes and after making this improvement + the improvement in this PR: https://github.com/cloudfoundry/gorouter/pull/174

[registry_perf.pdf](https://github.com/cloudfoundry/gorouter/files/1003783/registry_perf.pdf)